refactor: replace ListAPIs component with ListAPI for improved structure and organization
